+++ sentry-android-core/src/main/java/io/sentry/android/core/SentryAndroidOptions.java
@@ -94,13 +94,6 @@ public final class SentryAndroidOptions extends SentryOptions {
    */
   private boolean enableActivityLifecycleTracingAutoFinish = true;
 
-  /**
-   * Profiling traces rate. 101 hz means 101 traces in 1 second. Defaults to 101 to avoid possible
-   * lockstep sampling. More on
-   * https://stackoverflow.com/questions/45470758/what-is-lockstep-sampling
-   */
-  private int profilingTracesHz = 101;
-
   /** Interface that loads the debug images list */
   private @NotNull IDebugImagesLoader debugImagesLoader = NoOpDebugImagesLoader.getInstance();
 
@@ -364,22 +357,6 @@ public final class SentryAndroidOptions extends SentryOptions {
    */
   @Deprecated
   public void setProfilingTracesIntervalMillis(final int profilingTracesIntervalMillis) {}
-
-  /**
-   * Returns the rate the profiler will sample rates at. 100 hz means 100 traces in 1 second.
-   *
-   * @return Rate the profiler will sample rates at.
-   */
-  @ApiStatus.Internal
-  public int getProfilingTracesHz() {
-    return profilingTracesHz;
-  }
-
-  /** Sets the rate the profiler will sample rates at. 100 hz means 100 traces in 1 second. */
-  @ApiStatus.Internal
-  public void setProfilingTracesHz(final int profilingTracesHz) {
-    this.profilingTracesHz = profilingTracesHz;
-  }
 
   /**
    * Returns the Debug image loader